requirements.txt
streamlit
mysql-connector-python
pandas
python-dotenv
groq
